Moves AccountMapEntry types to submodule
