+++ accounts-db/src/accounts_index/in_mem_accounts_index.rs
@@ -1,8 +1,11 @@
 use {
     crate::{
         accounts_index::{
-            AccountMapEntry, AccountMapEntryInner, AccountMapEntryMeta, DiskIndexValue, IndexValue,
-            PreAllocatedAccountMapEntry, RefCount, SlotList, UpsertReclaim, ZeroLamport,
+            account_map_entry::{
+                AccountMapEntry, AccountMapEntryInner, AccountMapEntryMeta,
+                PreAllocatedAccountMapEntry,
+            },
+            DiskIndexValue, IndexValue, RefCount, SlotList, UpsertReclaim, ZeroLamport,
         },
         bucket_map_holder::{Age, AtomicAge, BucketMapHolder},
         bucket_map_holder_stats::BucketMapHolderStats,
